pi/requirements.txt
pyntcore>=2026.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
